[Bronze V] Title: 이상한 기호, Time: 208 ms, Memory: 17672 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -3,7 +3,7 @@ import java.util.*;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int a = sc.nextInt(), b = sc.nextInt();
+        long a = sc.nextLong(), b = sc.nextLong();
         System.out.println(((a + b)*(a - b)));
     }
 }
